rename producer contract file name and the namespace to be generic
