+++ src/main/kotlin/uk/gov/dluhc/registercheckerapi/mapper/CheckStatusMapper.kt
@@ -3,9 +3,9 @@ package uk.gov.dluhc.registercheckerapi.mapper
 import org.mapstruct.Mapper
 import org.mapstruct.MappingConstants
 import org.mapstruct.ValueMapping
+import uk.gov.dluhc.applicationsapi.messaging.models.RegisterCheckResult
 import uk.gov.dluhc.registercheckerapi.database.entity.CheckStatus
 import uk.gov.dluhc.registercheckerapi.dto.RegisterCheckStatus
-import uk.gov.dluhc.votercardapplicationsapi.messaging.models.RegisterCheckResult
 
 @Mapper
 interface CheckStatusMapper {
