+++ src/main/kotlin/uk/gov/dluhc/registercheckerapi/mapper/SourceTypeMapper.kt
@@ -5,10 +5,10 @@ import org.mapstruct.Mapper
 import org.mapstruct.ValueMapping
 import uk.gov.dluhc.registercheckerapi.dto.SourceType
 import uk.gov.dluhc.registercheckerapi.models.SourceSystem
+import uk.gov.dluhc.applicationsapi.messaging.models.SourceType as SourceTypeVcaSqsEnum
 import uk.gov.dluhc.registercheckerapi.database.entity.SourceType as SourceTypeEntityEnum
 import uk.gov.dluhc.registercheckerapi.dto.SourceType as SourceTypeDtoEnum
 import uk.gov.dluhc.registercheckerapi.messaging.models.SourceType as SourceTypeRcaSqsEnum
-import uk.gov.dluhc.votercardapplicationsapi.messaging.models.SourceType as SourceTypeVcaSqsEnum
 
 @Mapper
 interface SourceTypeMapper {
