+++ src/main/resources/openapi/sqs/applications-api-sqs-messaging-erop.yaml
