+++ src/test/kotlin/uk/gov/dluhc/registercheckerapi/mapper/SourceTypeMapperTest.kt
@@ -5,10 +5,10 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import uk.gov.dluhc.registercheckerapi.models.SourceSystem
+import uk.gov.dluhc.applicationsapi.messaging.models.SourceType as SourceTypeVcaSqsEnum
 import uk.gov.dluhc.registercheckerapi.database.entity.SourceType as SourceTypeEntityEnum
 import uk.gov.dluhc.registercheckerapi.dto.SourceType as SourceTypeDtoEnum
 import uk.gov.dluhc.registercheckerapi.messaging.models.SourceType as SourceTypeRcaSqsEnum
-import uk.gov.dluhc.votercardapplicationsapi.messaging.models.SourceType as SourceTypeVcaSqsEnum
 
 class SourceTypeMapperTest {
     private val mapper = SourceTypeMapperImpl()
